Modified Level 1 & Introduction Camera
I have changed the cameras to both be facing forwards. This means
changing rotation from 90 to 0. The canvas for the introduction level
was also changed from 90 to 0.

I understand why the camera was top down in level 1 but if we are able
to get the same behaviour of asteroids spawning in and coming in at the
new modified angle relative to the camera it would be easier moving
forwards.

I have also modified the 'far' distance on all the cameras so that they
cull off the back end of the planet, this is done for optimization
reasons. If we cannot see the back of the planet we don't need to render
it